+++ src/kgen/src/exprs/term_object.rs
@@ -30,7 +30,7 @@ impl<'ctx> TermObject {
         }
     }
 
-    pub fn eval(&self) -> u64 {
+    pub fn eval(&self) -> u32 {
         let mut first = self.first.eval();
         for val in &self.others {
             let (op, obj) = val;
