+++ src/kgen/src/jit/execute_function.rs
@@ -4,8 +4,8 @@ use inkwell::execution_engine::JitFunction;
 use super::super::program_object::CodeGen;
 use super::super::functions::function_object::function_parser;
 
-pub fn execute_function(text: &str, func_name: &str, arg: u64) -> Option<u64> {
-    type TestFunc = unsafe extern "C" fn(u64) -> u64;
+pub fn execute_function(text: &str, func_name: &str, arg: u32) -> Option<u32> {
+    type TestFunc = unsafe extern "C" fn(u32) -> u32;
 
     let context = &Context::create();
     let gen = CodeGen::new(context, "test");
@@ -18,7 +18,7 @@ pub fn execute_function(text: &str, func_name: &str, arg: u64) -> Option<u64> {
             let execution_engine = gen.module.create_jit_execution_engine(OptimizationLevel::None).unwrap();
 
             let func: JitFunction<TestFunc> = unsafe { execution_engine.get_function(func_name).unwrap() };
-            let ret: u64 = unsafe { func.call(arg) };
+            let ret: u32 = unsafe { func.call(arg) };
 
             Some(ret)
         },
